Reproducible case for anchor link encoding failures
Looking a little deeper into the source code of the failures, I noticed that not _all_ the non-ASCII links were failing, only some of them. The site gets built with relative links between pages, and it was links to another page with a "../page/#hash" syntax which failed.

This commit adds a test directory which mimics that structure and reliably reproduces the error... I think!